+++ mediamanager.php
@@ -14,7 +14,7 @@
  * @wordpress-plugin
  * Plugin Name: Media Manager
  * Description: Virtual folder organization and smart management for the WordPress Media Library.
- * Version: 0.1.14
+ * Version: 0.1.15
  * Requires at least: 6.8
  * Requires PHP: 8.3
  * Author: Per Soderlind